Combine Drone into Config class
