+++ lib/drone_chef.rb
@@ -2,7 +2,6 @@
 # Drone plugin for managing Chef publishing/deployment of cookbooks
 #
 
-require 'drone_chef/drone'
 require 'drone_chef/config'
 require 'drone_chef/chef_server'
 require 'drone_chef/supermarket'
